Move clone() to CloningMixin
